requirements.txt
# FastAPI 核心
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Pydantic
pydantic>=2.10.0
pydantic-settings>=2.6.0
pydantic[email]>=2.10.0

# 配置管理
python-dotenv>=1.0.0

# MongoDB (Motor + Beanie ODM)
motor>=3.6.0
beanie>=1.27.0

# Redis
redis>=5.2.0

# JWT
python-jose[cryptography]>=3.3.0

# 密码哈希
passlib[bcrypt]>=1.7.4